+++ tests/datasets/test_structure.py
@@ -262,6 +262,29 @@ def test_import_property_string_enum_without_prepare_uses_source_value():
     assert enum_item_2.prepare == '"two"'
 
 
+def test_import_property_string_enum_without_quoted_prepare_adds_error():
+    manifest = (
+        "id,dataset,resource,base,model,property,type,ref,source,prepare,level,access,uri,title,description\n"
+        ",example,,,,,,,,,,,,,\n"
+        ",,,,Dataset,,,id,,,,,,,\n"
+        ",,,,,id,integer,,,,,,,,\n"
+        ",,,,,str_enum,string,,STR_ENUM,,,,,,\n"
+        ",,,,,,enum,,one,one,,,,,\n"
+    )
+    reader = csv.DictReader(io.StringIO(manifest))
+    model = "example/Dataset"
+
+    state = read(reader)
+    assert state.errors == []
+
+    str_enum_property = state.manifest.models[model].properties["str_enum"]
+    assert str_enum_property.type == "string"
+    enum_item_1 = str_enum_property.enums[""][0]
+    assert enum_item_1.source == "one"
+    assert enum_item_1.prepare == "one"
+    assert enum_item_1.errors == ['Reikšmė "one" turi būti string tipo.']
+
+
 def test_import_property_not_string_enum_without_prepare_results_in_error():
     manifest = (
         "id,dataset,resource,base,model,property,type,ref,source,prepare,level,access,uri,title,description\n"
@@ -284,11 +307,48 @@ def test_import_property_not_string_enum_without_prepare_results_in_error():
     enum_item_1 = int_enum_property.enums[""][0]
     assert enum_item_1.source == "1"
     assert enum_item_1.prepare == ""
-    assert enum_item_1.errors == ['Duomenų reikšmė (source: "1") privalo turėti nurodytą "prepare" stulpelį.']
+    assert enum_item_1.errors == [
+        'Duomenų reikšmė (source: "1") privalo turėti nurodytą "prepare" stulpelį.',
+        'Reikšmė "" turi būti integer tipo.',
+    ]
     enum_item_2 = int_enum_property.enums[""][1]
     assert enum_item_2.source == "2"
     assert enum_item_2.prepare == ""
     assert enum_item_2.errors == [
         'Duomenų reikšmė (source: "2") privalo turėti nurodytą "prepare" stulpelį.',
-        'Galima reikšmė (source: "2") "" jau egzistuoja.',
+        'Reikšmė "" turi būti integer tipo.',
     ]
+
+
+def test_import_enum_checks_uniqueness_based_on_source_and_prepare():
+    manifest = (
+        "id,dataset,resource,base,model,property,type,ref,source,prepare,level,access,uri,title,description\n"
+        ",example,,,,,,,,,,,,,\n"
+        ",,,,Dataset,,,id,,,,,,,\n"
+        ",,,,,id,integer,,,,,,,,\n"
+        ",,,,,int_enum,integer,,INT_ENUM,,,,,,\n"
+        ",,,,,,enum,,1,1,,,,,\n"
+        ",,,,,,,,2,1,,,,,\n"
+        ",,,,,,,,1,1,,,,,\n"
+    )
+    reader = csv.DictReader(io.StringIO(manifest))
+    model = "example/Dataset"
+
+    state = read(reader)
+    assert state.errors == []
+
+    int_enum_property = state.manifest.models[model].properties["int_enum"]
+    assert int_enum_property.type == "integer"
+
+    enum_item_1 = int_enum_property.enums[""][0]
+    assert enum_item_1.source == "1"
+    assert enum_item_1.prepare == "1"
+    assert enum_item_1.errors == []
+    enum_item_2 = int_enum_property.enums[""][1]
+    assert enum_item_2.source == "2"
+    assert enum_item_2.prepare == "1"
+    assert enum_item_2.errors == []
+    enum_item_3 = int_enum_property.enums[""][2]
+    assert enum_item_3.source == "1"
+    assert enum_item_3.prepare == "1"
+    assert enum_item_3.errors == ['Galima reikšmė (source: "1") "1" jau egzistuoja.']