+++ vitrina/structure/services.py
@@ -259,6 +259,7 @@ def _load_enums(
                 if en := existing_enum_items.filter(
                     metadata__content_type=enum_ct,
                     metadata__name=meta.name,
+                    metadata__source=meta.source,
                     metadata__prepare=meta.prepare,
                     metadata__dataset=dataset,
                 ).first():
@@ -1417,6 +1418,7 @@ def _enums_to_tabular(obj: models.Model, separator: bool = False, version: Versi
                         "ref": enum.name if first else "",
                         "source": meta.source,
                         "prepare": meta.prepare,
+                        "level": meta.level_given,
                         "access": _get_access(meta.access),
                         "title": meta.title,
                         "description": meta.description,