+++ vitrina/structure/utils.py
@@ -3,7 +3,12 @@ from typing import TypeVar, Generic, Any
 
 from django.utils.translation import gettext_lazy as _
 
+from vitrina.structure.helpers import is_quoted
+
 T = TypeVar("T")
+
+
+VALID_BOOLEAN_PREPARE_VALUES = ("true", "false")
 
 
 class TypeCheckerError(Exception):
@@ -15,7 +20,10 @@ class TypeChecker(ABC, Generic[T]):
 
 
 class StringTypeChecker(TypeChecker[str]):
-    pass
+    def check_enum_item_value(self, value: str) -> None:
+        if not is_quoted(value):
+            error_msg = _('Reikšmė "{value}" turi būti string tipo.').format(value=value)
+            raise TypeCheckerError(error_msg)
 
 
 class IntegerTypeChecker(TypeChecker[int]):
@@ -23,7 +31,18 @@ class IntegerTypeChecker(TypeChecker[int]):
         try:
             int(value)
         except (TypeError, ValueError):
-            raise TypeCheckerError(_(f'Reikšmė "{value}" turi būti integer tipo.'))
+            error_msg = _('Reikšmė "{value}" turi būti integer tipo.').format(value=value)
+            raise TypeCheckerError(error_msg)
+
+
+class BooleanTypeChecker(TypeChecker[bool]):
+    def check_enum_item_value(self, value: str) -> None:
+        if value not in VALID_BOOLEAN_PREPARE_VALUES:
+            possible_values = ", ".join(VALID_BOOLEAN_PREPARE_VALUES)
+            error_msg = _('Reikšmė "{value}" turi būti boolean tipo. Viena iš: {possible_values}').format(
+                value=value, possible_values=possible_values
+            )
+            raise TypeCheckerError(error_msg)
 
 
 class NotImplementedTypeChecker(TypeChecker[Any]):
@@ -34,6 +53,7 @@ class NotImplementedTypeChecker(TypeChecker[Any]):
 TYPE_CHECKER_MAP = {
     "string": StringTypeChecker(),
     "integer": IntegerTypeChecker(),
+    "boolean": BooleanTypeChecker(),
 }